radio_probe: silence -Wformat on uint32_t rev
On the ESP32 xtensa/riscv toolchain uint32_t is unsigned long, so the %u format specifier for the ZNP SYS_VERSION revision word triggers -Wformat. Cast to unsigned (matches the style already used for payload_len and PROBE_TIMEOUT_MS in the same file) at both the ESP_LOGD and snprintf call sites.

No behavioural change - the value still fits in 32 bits and prints identically.

+++ components/radio_probe/znp_probe.cpp
@@ -109,9 +109,9 @@ bool RadioProbe::probe_znp_(std::string &result) {
                              (static_cast<uint32_t>(buf[11]) << 16) |
                              (static_cast<uint32_t>(buf[12]) << 24);
         ESP_LOGD(TAG, "ZNP SYS_VERSION: transport=%u product=%u %u.%u.%u rev=%u", buf[4], buf[5],
-                 buf[6], buf[7], buf[8], rev);
+                 buf[6], buf[7], buf[8], static_cast<unsigned>(rev));
         char out[16];
-        snprintf(out, sizeof(out), "%u", rev);
+        snprintf(out, sizeof(out), "%u", static_cast<unsigned>(rev));
         result = out;
         return true;
       }
